+++ application/src/main/java/org/opentripplanner/graph_builder/GraphBuilder.java
@@ -17,6 +17,7 @@ import org.opentripplanner.ext.empiricaldelay.EmpiricalDelayRepository;
 import org.opentripplanner.ext.stopconsolidation.StopConsolidationRepository;
 import org.opentripplanner.framework.application.OTPFeature;
 import org.opentripplanner.framework.application.OtpAppException;
+import org.opentripplanner.gbfs.network.GbfsNetworkOverrides;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssueStore;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssueSummary;
 import org.opentripplanner.graph_builder.model.GraphBuilderModule;
@@ -27,6 +28,7 @@ import org.opentripplanner.routing.fares.FareServiceFactory;
 import org.opentripplanner.service.osminfo.OsmInfoGraphBuildRepository;
 import org.opentripplanner.service.streetdetails.StreetDetailsRepository;
 import org.opentripplanner.service.vehicleparking.VehicleParkingRepository;
+import org.opentripplanner.service.vehiclerental.VehicleRentalRepository;
 import org.opentripplanner.service.worldenvelope.WorldEnvelopeRepository;
 import org.opentripplanner.standalone.config.BuildConfig;
 import org.opentripplanner.street.StreetRepository;
@@ -78,6 +80,7 @@ public class GraphBuilder implements Runnable {
    */
   public static GraphBuilder create(
     BuildConfig config,
+    GbfsNetworkOverrides gbfsNetworkOverrides,
     GraphBuilderDataSources dataSources,
     Graph graph,
     OsmInfoGraphBuildRepository osmInfoGraphBuildRepository,
@@ -88,6 +91,7 @@ public class GraphBuilder implements Runnable {
     TransferRepository transferRepository,
     WorldEnvelopeRepository worldEnvelopeRepository,
     VehicleParkingRepository vehicleParkingService,
+    VehicleRentalRepository vehicleRentalRepository,
     @Nullable EmissionRepository emissionRepository,
     @Nullable EmpiricalDelayRepository empiricalDelayRepository,
     @Nullable StopConsolidationRepository stopConsolidationRepository,
@@ -104,6 +108,7 @@ public class GraphBuilder implements Runnable {
     GraphBuilderFactory.Builder builder = DaggerGraphBuilderFactory.builder();
     builder
       .config(config)
+      .gbfsNetworkOverrides(gbfsNetworkOverrides)
       .graph(graph)
       .osmInfoGraphBuildRepository(osmInfoGraphBuildRepository)
       .streetDetailsRepository(streetDetailsRepository)
@@ -112,6 +117,7 @@ public class GraphBuilder implements Runnable {
       .transferRepository(transferRepository)
       .worldEnvelopeRepository(worldEnvelopeRepository)
       .vehicleParkingRepository(vehicleParkingService)
+      .vehicleRentalRepository(vehicleRentalRepository)
       .stopConsolidationRepository(stopConsolidationRepository)
       .emissionRepository(emissionRepository)
       .empiricalDelayRepository(empiricalDelayRepository)
@@ -192,6 +198,7 @@ public class GraphBuilder implements Runnable {
     }
 
     if (loadStreetGraph || hasOsm) {
+      graphBuilder.addModuleOptional(factory.vehicleRentalGeofencingGraphBuilder());
       graphBuilder.addModule(factory.graphCoherencyCheckerModule());
     }