+++ application/src/main/java/org/opentripplanner/standalone/configure/ConstructApplicationFactory.java
@@ -52,7 +52,6 @@ import org.opentripplanner.service.vehicleparking.VehicleParkingService;
 import org.opentripplanner.service.vehicleparking.configure.VehicleParkingServiceModule;
 import org.opentripplanner.service.vehiclerental.VehicleRentalRepository;
 import org.opentripplanner.service.vehiclerental.VehicleRentalService;
-import org.opentripplanner.service.vehiclerental.configure.VehicleRentalRepositoryModule;
 import org.opentripplanner.service.vehiclerental.configure.VehicleRentalServiceModule;
 import org.opentripplanner.service.worldenvelope.WorldEnvelopeRepository;
 import org.opentripplanner.service.worldenvelope.WorldEnvelopeService;
@@ -105,7 +104,6 @@ import org.opentripplanner.warmup.configure.WarmupModule;
     TransitModule.class,
     TransferServiceModule.class,
     VehicleParkingServiceModule.class,
-    VehicleRentalRepositoryModule.class,
     VehicleRentalServiceModule.class,
     ViaModule.class,
     WarmupModule.class,
@@ -216,6 +214,9 @@ public interface ConstructApplicationFactory {
 
     @BindsInstance
     Builder vehicleParkingRepository(VehicleParkingRepository parkingRepository);
+
+    @BindsInstance
+    Builder vehicleRentalRepository(VehicleRentalRepository rentalRepository);
 
     @BindsInstance
     Builder dataImportIssueSummary(DataImportIssueSummary issueSummary);